--- .pytest_cache/v/cache/lastfailed
@@ -1,83 +1,0 @@
-{
-  "tests/test_app.py::test_invalid_url": true,
-  "tests/test_app.py::test_method_not_allowed": true,
-  "tests/test_app.py::test_welcome_message": true,
-  "tests/test_create_incident.py::test_admin_cannot_create_a_red_flag": true,
-  "tests/test_create_incident.py::test_create_a_red_flag_with_bad_json_format_data": true,
-  "tests/test_create_incident.py::test_create_a_red_flag_with_invalid_token": true,
-  "tests/test_create_incident.py::test_create_a_red_flag_with_valid_data": true,
-  "tests/test_create_incident.py::test_create_a_red_flag_without_a_token": true,
-  "tests/test_create_incident.py::test_create_a_red_flag_without_data": true,
-  "tests/test_create_incident.py::test_create_a_red_flag_without_wrong_input": true,
-  "tests/test_create_incident.py::test_create_an_intervention_with_valid_data": true,
-  "tests/test_delete_incident.py::test_delete_an_intervention": true,
-  "tests/test_delete_incident.py::test_delete_red_flag": true,
-  "tests/test_delete_incident.py::test_delete_red_flag_for_another_user": true,
-  "tests/test_delete_incident.py::test_delete_red_flag_for_with_status_other_than_draft": true,
-  "tests/test_delete_incident.py::test_delete_red_flag_with_invalid_format_red_flag_id": true,
-  "tests/test_delete_incident.py::test_delete_red_flag_with_red_flag_id_which_does_not_exist": true,
-  "tests/test_delete_incident.py::test_delete_red_flag_without_a_access_token": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_created_by_another_user": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_created_by_the_current_user": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_for_a_intervention_record_which_does_not_exist": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_for_a_intervention_record_with_without_a_comment": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_with_an_invalid_intervention_id": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_with_status_other_than_draft": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_without_a_token": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_without_comment_data": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_created_by_another_user": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_created_by_the_current_user": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_for_a_red_flag_record_which_does_not_exist": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_for_a_red_flag_record_with_without_a_comment": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_with_an_invalid_red_flag_id": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_with_status_other_than_draft": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_without_a_token": true,
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_without_comment_data": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_status_of_draft": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_status_other_than_draft": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_which_does_not_belong_to_user": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_which_does_not_exist": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_with_invalid_intervention_id": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_with_invalid_location_coordinates": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_without_a_token": true,
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_without_location_data": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_status_of_draft": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_status_other_than_draft": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_which_does_not_belong_to_user": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_which_does_not_exist": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_with_invalid_location_coordinates": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_with_invalid_red_flag_id": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_without_a_token": true,
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_without_location_data": true,
-  "tests/test_edit_incident_status.py::test_edit_a_intervention_status": true,
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status": true,
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_for_a_red_flag_record_which_does_not_exist": true,
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_with_an_invalid_red_flag_id": true,
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_with_invalid_status_data": true,
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_without_a_token": true,
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_without_request_data": true,
-  "tests/test_edit_incident_status.py::test_non_admin_edit_a_red_flag_status": true,
-  "tests/test_get_incidents.py::test_admin_get_a_intervention": true,
-  "tests/test_get_incidents.py::test_admin_get_a_red_flag": true,
-  "tests/test_get_incidents.py::test_get_a_intervention": true,
-  "tests/test_get_incidents.py::test_get_a_intervention_for_another_user": true,
-  "tests/test_get_incidents.py::test_get_a_red_flag": true,
-  "tests/test_get_incidents.py::test_get_a_red_flag_for_another_user": true,
-  "tests/test_get_incidents.py::test_get_a_red_flag_with_id_which_does_not_exist": true,
-  "tests/test_get_incidents.py::test_get_all_interventions": true,
-  "tests/test_get_incidents.py::test_get_all_interventions_with_expired_token": true,
-  "tests/test_get_incidents.py::test_get_all_interventions_without_token": true,
-  "tests/test_get_incidents.py::test_get_all_red_flags": true,
-  "tests/test_get_incidents.py::test_get_all_red_flags_with_expired_token": true,
-  "tests/test_get_incidents.py::test_get_all_red_flags_without_token": true,
-  "tests/test_login.py::test_user_login_with_correct_credentials_data": true,
-  "tests/test_login.py::test_user_login_with_valid_credentials_": true,
-  "tests/test_login.py::test_user_login_with_wrong_credentials_": true,
-  "tests/test_login.py::test_user_login_with_wrong_keys_in_credentials_data": true,
-  "tests/test_login.py::test_user_login_without_data_": true,
-  "tests/test_signup.py::test_register_user_with_duplicate_data": true,
-  "tests/test_signup.py::test_register_with_no_data": true,
-  "tests/test_signup.py::test_register_with_valid_format_data": true,
-  "tests/test_signup.py::test_register_with_wrong_invalid_format_data": true,
-  "tests/test_signup.py::test_register_with_wrong_key": true
-}
