--- .pytest_cache/v/cache/nodeids
@@ -1,83 +1,0 @@
-[
-  "tests/test_app.py::test_invalid_url",
-  "tests/test_app.py::test_method_not_allowed",
-  "tests/test_app.py::test_welcome_message",
-  "tests/test_create_incident.py::test_create_a_red_flag_without_a_token",
-  "tests/test_create_incident.py::test_create_a_red_flag_with_invalid_token",
-  "tests/test_create_incident.py::test_create_a_red_flag_with_bad_json_format_data",
-  "tests/test_create_incident.py::test_admin_cannot_create_a_red_flag",
-  "tests/test_create_incident.py::test_create_a_red_flag_without_data",
-  "tests/test_create_incident.py::test_create_a_red_flag_with_valid_data",
-  "tests/test_create_incident.py::test_create_a_red_flag_without_wrong_input",
-  "tests/test_create_incident.py::test_create_an_intervention_with_valid_data",
-  "tests/test_delete_incident.py::test_delete_red_flag_without_a_access_token",
-  "tests/test_delete_incident.py::test_delete_red_flag_with_red_flag_id_which_does_not_exist",
-  "tests/test_delete_incident.py::test_delete_red_flag_with_invalid_format_red_flag_id",
-  "tests/test_delete_incident.py::test_delete_red_flag_for_another_user",
-  "tests/test_delete_incident.py::test_delete_red_flag_for_with_status_other_than_draft",
-  "tests/test_delete_incident.py::test_delete_red_flag",
-  "tests/test_delete_incident.py::test_delete_an_intervention",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_without_a_token",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_with_an_invalid_red_flag_id",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_without_comment_data",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_for_a_red_flag_record_which_does_not_exist",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_for_a_red_flag_record_with_without_a_comment",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_created_by_another_user",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_created_by_the_current_user",
-  "tests/test_edit_incident_comment.py::test_edit_a_red_flag_comment_with_status_other_than_draft",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_without_a_token",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_with_an_invalid_intervention_id",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_without_comment_data",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_for_a_intervention_record_which_does_not_exist",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_for_a_intervention_record_with_without_a_comment",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_created_by_another_user",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_created_by_the_current_user",
-  "tests/test_edit_incident_comment.py::test_edit_a_intervention_comment_with_status_other_than_draft",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_without_a_token",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_without_location_data",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_which_does_not_exist",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_which_does_not_belong_to_user",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_with_invalid_red_flag_id",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_status_of_draft",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_with_invalid_location_coordinates",
-  "tests/test_edit_incident_location.py::test_edit_a_red_flag_location_status_other_than_draft",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_without_a_token",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_without_location_data",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_which_does_not_exist",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_which_does_not_belong_to_user",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_with_invalid_intervention_id",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_status_of_draft",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_with_invalid_location_coordinates",
-  "tests/test_edit_incident_location.py::test_edit_a_intervention_location_status_other_than_draft",
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_without_a_token",
-  "tests/test_edit_incident_status.py::test_non_admin_edit_a_red_flag_status",
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_with_an_invalid_red_flag_id",
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_without_request_data",
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_with_invalid_status_data",
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status_for_a_red_flag_record_which_does_not_exist",
-  "tests/test_edit_incident_status.py::test_edit_a_red_flag_status",
-  "tests/test_edit_incident_status.py::test_edit_a_intervention_status",
-  "tests/test_get_incidents.py::test_get_all_red_flags_without_token",
-  "tests/test_get_incidents.py::test_get_all_red_flags_with_expired_token",
-  "tests/test_get_incidents.py::test_get_all_red_flags",
-  "tests/test_get_incidents.py::test_get_a_red_flag",
-  "tests/test_get_incidents.py::test_get_a_red_flag_with_id_which_does_not_exist",
-  "tests/test_get_incidents.py::test_get_a_red_flag_for_another_user",
-  "tests/test_get_incidents.py::test_admin_get_a_red_flag",
-  "tests/test_get_incidents.py::test_get_all_interventions_without_token",
-  "tests/test_get_incidents.py::test_get_all_interventions_with_expired_token",
-  "tests/test_get_incidents.py::test_get_all_interventions",
-  "tests/test_get_incidents.py::test_get_a_intervention",
-  "tests/test_get_incidents.py::test_get_a_intervention_for_another_user",
-  "tests/test_get_incidents.py::test_admin_get_a_intervention",
-  "tests/test_login.py::test_user_login_with_valid_credentials_",
-  "tests/test_login.py::test_user_login_without_data_",
-  "tests/test_login.py::test_user_login_with_wrong_credentials_",
-  "tests/test_login.py::test_user_login_with_wrong_keys_in_credentials_data",
-  "tests/test_login.py::test_user_login_with_correct_credentials_data",
-  "tests/test_signup.py::test_register_with_no_data",
-  "tests/test_signup.py::test_register_with_wrong_key",
-  "tests/test_signup.py::test_register_with_wrong_invalid_format_data",
-  "tests/test_signup.py::test_register_with_valid_format_data",
-  "tests/test_signup.py::test_register_user_with_duplicate_data"
-]
